feat: Enhance authentication flow by integrating JWT service for session management and improving cookie handling

+++ internal/delivery/dto/auth_dto.go
@@ -18,5 +18,5 @@ type LoginResponseDTO struct {
 
 type RefreshResponseDTO struct {
 	AccessToken string `json:"accessToken"`
-	ExpiresIn   int64  `json:"expiresIn"`
+	ExpiresIn   int    `json:"expiresIn"`
 }
